Install akonadi-server.service, akonadi.target and background-akonadi.slice
akonadi-server.service is a systemd service that runs the Akonadi Server.
THe akonadi.target unit is a target that enables DBus activation for Akonadi
and encompases all the Akonadi-related services.
Finally, background-akonadi.slice is a slice (nested in the background.slice)
which creates a separate cgroup for Akonadi. This will, amongst other things,
make it possible to impose resource limiting on Akonadi.