Akonadi Control: clean-up retired resources on start
When during start, Akonadi Control finds configuration for resources whose
agent types are missing (they are no longer installed) and are present on
a list of retired agent types, it can safely forget about them and remove
all their left-over data from Akonadi.

The list contains resources that we are planning to retire in the near
future. When users fail to migrate away from them manually and update to
a new version of Akonadi that will no longer have those resources, they
would be forever stuck with their Collections and Items in the database
with no easy way to remove them. This code will take care of cleaning up
the mess for them.

We have an explicit list of retired agent types, so that we don't
accidentally wipe data of still-living resources in case the user
installation gets corrupted, there's a packaging bug that causes the
resource to be missing from the package and other even more unlikely
scenarios...but it's better to be careful in this case.