QueryBuilder: make it possible to use QSet in WHERE IN queries
This allows us to avoid one entire iteration over the QSet to construct
a QList from it. The implementation is still not perfect (performance-wise),
but right now that's still a good trade-off between performance and ease-
of-use of the QueryBuilder. Proper generic implementation (in order to
support QLists and QSets of multiple types than just qint64) would require
template magic and bigger changes to the QueryBuilder itself - I'm leaving
that up for future Dan to figure out...