QueryBuilder: allow inserting multiple rows with single INSERT query
Using QueryBuilder::setColumnValues it's possible to specify a list of values
to insert into each column. This allows inserting multiple rows in a single
query, which replaces the need of a transaction and multiple inserts, usually
in cases like creating relations between entities.

The implementation is sub-optimal, because we store data as a list of values
per column, while for the query we need them as list of values per row.
This should be optimized in the future, especially if we start using this API
more, but it would require larger changes to the QueryBuilder.