Completely remove Akonadi Relations (ABI/API break)
This has only ever worked for Kolab resource, no-one else
can sync this kind of information. Long-term we want to
keep this kind of information in the search index rather
than directly in Akonadi database, from where it cannot
be reliably synced, especially in case of cross-resource
relations.

Since we don't officially maintain public stable API/ABI
as of yet, I decided to remove the feature completely
instead of just deprecating everything and keeping it
around until KPim7. If you are a dev, you may need to
rebuild most of your PIM stack.