Fetch Tag GID by default when fetching Item Tags
When ItemFetchScope.fetchTags was set to true, it would only
fetch tag IDs by default. However, that's rarely useful, majority
of the code needs to enable additional parameters in the
ItemFetchScope.tagFetchScope (and also be aware that it needs to
disable fetchOnlyId, which is counter-intuitive). With this change,
just enabling itemFetchScope.fetchTags, the caller will by default
receive useful Tag data without having to configure anything else.