Port away from deprecated QSqlQuery copy constructor
QSqlQuery should be move-only. This is mostly OK, as most of our code was used to
obtain a copy from the QueryBuilder and then discard the QueryBuilder. Instead we
now get a reference to the instance owned by the QueryBuilder and keep the query
builder around for as long as we need the QSqlQuery. This has a nice side-effect
that I could remove explicit calls to QSqlQuery::finish() in most places and we
instead call it in the QueryBuilder destructor, since it rarely lives much longer
than necessary.

The only problem was the QueryCache, which relied on handing our copies of the
original QSqlQuery. But the cache is per-thread and we should never have two
overlapping identical querties in a single thread, so this change modifies the
cache move the query to the caller (QueryBuilder) and forget about it, and the
QueryBuilder will insert the query back into the cache once it is done with it.
This also retains the LRU nature of the cache.

The only loss of functionality is in the StorageDebugger which used to
make a copy of the query and extract results from it (and thus not
affecting the original query). Since we can no longe make a copy of the
query, we can no longer extract the results. But since it only affects a
rarely used debug feature, we don't care that much.

Finally, this change restructures, cleans up and modernizes the QueryBuilderTest,
with each test data being scoped and not reusing QueryBuilders, bind values or
anything else from previous test data.

Finally, thanks to this change, we can compile Akonadi with
QT_DISABLE_DEPRECATED_BEFORE 6.2.