ItemSync: use DB transactions to optimize sync speed
Creating one transaction for each new item is costly. Instead we open a
transaction on the first item and commit it automatically either when
the sync finishes, or after at most 100 ms. This ensures that the
transaction never stays open for too long, while we can still group
insertion or update of several Items into a single database transaction.