Group ItemsTagsChanged notification on tag deletion by collection
The notification collector assumes all changed Items belong to the same
collection, since that's historically how this usually works. In case of
deleting a tag, the affected Items can belong to many different collections.
To make sure that the notification batch is processed correctly by all
recipients, we split it to one notification per collection.