Send entire Tags in ItemsTagsChanged notification
Previously we were only sending Tag ID and relying on the Monitor's
EntityCache to either have the Tag cached or to retrieve it from Akonadi.
This worked just fine, except for one particular not-so-corner case: when
a Tag is deleted, we send an ItemsTagsChanged notification about all tagged
Items to remove the tag. The problem is that by the time the notification
has been received by clients, the Tag has already been deleted from the
Akonadi database, and so the EntityCache in Monitor fails to fetch it,
effectively turning the ItemsTagsChanged notification into a no-op.

To solve this problem, we need to send the entire Tag as part of the
notification, so that the Monitor can fully reconstruct the Tag from
notification without having to fetch it from Akonadi.