collectioncombobox.cpp - add a connect to deal with the combobox changes
Don't ignore the combobox signal indicating that the selection index
has changed. We'll emit the currentChanged signal in that event.

Then users of this widget can connect to that currentChanged signal
and do something useful.

Specifically in the case of the incidenceeditor dialog, we can
enable/disable the Apply button.

BUG: 498551


(cherry picked from commit 9305dcf7fcbde4accf11fd78be95077144ac4355)

Co-authored-by: Allen Winter <winter@kde.org>