Only consider the server as ready when all agents are on DBus
It goes like this: AgentManager considers all agents ready ->
AgentManager registers control interface -> ServerManager
(core/servermanager.cpp) reports server ready when it sees the
control interface appear on the bus -> clients consider the server
ready and operate as if, well, everything was ready.

Since clients generally talk to agents via DBus, reporting
"everything is ready" before it's possible to talk to all agents
can cause some messy misunderstandings such as KMail or the
mail filter agent loading filters and throwing errors (with
currently pretty bad consequences such as *unconditionally
deleting* affected filters) because they don't see the
corresponding resources.

It is not clear that "interfaces registered on DBus" is even
enough, it also depends on how agents respond to early requests
that they could be reasonably expected to handle, but fail to - and
what the clients of these agents expect.
E.g. hypothetically, asking a just launched IMAP agent for the
server address could be reasonably expected to work, but asking it
to fetch some items could not be because it's not even logged in.
At some point, it is not the server's problem anymore, but I think
what this commit handles *is* still the server's problem.