widgets/standardactionmanager: Filter clipboard change events
Under Wayland, QClipboard::changed signal is triggered very often, for
example each time a QMenu is opened, or when the focus of the main window
changed.
This is an issue because Actions (and associated QMenu) are updated for
various events, including the clipboard data change event. And when
updating Actions, the QMenu are updated: old sub-menu is destroyed which
close the sub-menu...

To fix that only update the actions when the clipboard truly changed.

BUG: 485661,508727


(cherry picked from commit 5c1aba602aa461dcf06dfe8b23f52df074b3f35c)

Co-authored-by: Benjamin Robin <benjamin.robin@bootlin.com>