Prefer mariaDB to Oracle MySQL
Linux distributions are starting to deprecate mysqld in favor of mariadb

Therefore, at runtime we search for the mariadb commands first and
fallback to mysqld.

Nothing else really changes in here except which commands are invoked.
ie, the mysql.socket, mysql.conf, etc files are still used and handled
as before.