Allow to retrieve const shared pointer item payloads
For value types this just works by assigning the result of Item::payload
e.g. to a const auto variable or wrapping is in std::as_const. For const-
correct use of types inside a shared pointer we however need the const
pointed-to type. To make that more convenient, allow to retrieve those
directly and apply the necessary conversion internally.

This is particularly relevant for KMime::Message payloads, as those have
different non-mutating const overloads.