Always use flush() instead of waitForBytesWritten()
waitForBytesWritten() does one really useful thing, it polls until
the timeout expires, but it only guarantees to write at least one
byte, which isn't all that much better than flush(), which doesn't
retry at all. Both don't guarantee that everything is written, so
where flush() fails, waitForBytesWritten() usually isn't guaranteed
to work neither. In practice, both rely on writing a small amount
of data to a much larger and relatively empty buffer in the OS (or
an event loop to keep writing the rest).
The disadvantage of waitForBytesWritten() is that it can emit
readyRead(), which (in connected slots) can do... many things...
that are possibly not expected to happen at that point.

This fixes at least one known bug in NotificationSubscriber where
something in a readyRead() handler flushes the write buffer, so
the later attempt in waitForBytesWritten() to do the same fails
and it (wrongly) returns an error.
I am working on making that not an error in Qt, but in the
meantime, this should also work.