Fix (somewhat) spuriously failing waitForBytesWritten()
waitForBytesWritten() emits readyRead() (if there is data
available) even before trying to write. If a slot connected to that
signal somehow flushes the write buffer, waitForBytesWritten()
later also tries to flush it and fails... slowly, after its
timeout.
To work around that, call flush() instead of waitForBytesWritten().
flush() really only writes out the write buffer. If that is a
problem here, it's at least less bad than the original problem of
waitForBytesWritten() failing with a timeout error.